face-app/requirements.txt
# Autor: Álvaro Muñoz Panadero - alvaromp.dev@gmail.com
# Face Recognition System

# Core dependencies
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.1
scikit-learn==1.3.0

# API dependencies
fastapi==0.103.2
uvicorn==0.23.2
python-multipart==0.0.6

# Optional: For development and testing
requests==2.31.0